@@ -1,6 +1,6 @@
 {
   "name": "commander",
-  "version": "2.20.3",
+  "version": "4.1.1",
   "description": "the complete solution for node.js command-line programs",
   "keywords": [
     "commander",
@@ -15,8 +15,8 @@
     "url": "https://github.com/tj/commander.js.git"
   },
   "scripts": {
-    "lint": "eslint index.js",
-    "test": "node test/run.js && npm run test-typings",
+    "lint": "eslint index.js \"tests/**/*.js\"",
+    "test": "jest && npm run test-typings",
     "test-typings": "tsc -p tsconfig.json"
   },
   "main": "index",
@@ -26,13 +26,16 @@
   ],
   "dependencies": {},
   "devDependencies": {
-    "@types/node": "^12.7.8",
-    "eslint": "^6.4.0",
-    "should": "^13.2.3",
-    "sinon": "^7.5.0",
+    "@types/jest": "^24.0.23",
+    "@types/node": "^12.12.11",
+    "eslint": "^6.7.0",
+    "eslint-plugin-jest": "^22.21.0",
+    "jest": "^24.8.0",
     "standard": "^14.3.1",
-    "ts-node": "^8.4.1",
-    "typescript": "^3.6.3"
+    "typescript": "^3.7.2"
   },
-  "typings": "typings/index.d.ts"
+  "typings": "typings/index.d.ts",
+  "engines": {
+    "node": ">= 6"
+  }
 }
